# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mock API",
    "version": "1.0.0",
    "description": "Categories va Products CRUD API - soxta ma'lumotlar bilan."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/categories": {
      "get": {
        "summary": "Get all categories",
        "responses": {
          "200": {
            "description": "List of categories"
          }
        }
      },
      "post": {
        "summary": "Create new category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created"
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "Get category by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Category"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated"
          }
        }
      },
      "delete": {
        "summary": "Delete category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted"
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "List of products"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get product by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/categories/{cid}/products": {
      "get": {
        "summary": "Get all products in category",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products list"
          }
        }
      },
      "post": {
        "summary": "Create product in category",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created"
          }
        }
      }
    },
    "/categories/{cid}/products/{pid}": {
      "put": {
        "summary": "Update product in category",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated"
          }
        }
      },
      "delete": {
        "summary": "Delete product",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          }
        },
        "required": [
          "title"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "badge": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "rating": {
            "type": "integer"
          },
          "weight": {
            "type": "integer"
          },
          "basePrice": {
            "type": "number"
          },
          "sizes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "doughs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "ingredients": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "extras": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "inCart": {
            "type": "boolean"
          },
          "quantity": {
            "type": "integer"
          }
        },
        "required": [
          "title"
        ]
      }
    }
  }
}